+++ test/remix/client-entry.test.ts
@@ -28,17 +28,17 @@ describe('initializeSentryOnEntryClient', () => {
         useMatches,
       } from "@remix-run/react";
 
-      import * as Sentry from "@sentry/remix";
+      import {  init, replayIntegration, browserTracingIntegration,} from "@sentry/remix";
 
-      Sentry.init({
+      init({
           dsn: "https://sentry.io/123",
           tracesSampleRate: 1,
 
-          integrations: [Sentry.browserTracingIntegration({
+          integrations: [browserTracingIntegration({
             useEffect,
             useLocation,
             useMatches
-          }), Sentry.replayIntegration({
+          }), replayIntegration({
               maskAllText: true,
               blockAllMedia: true
           })],
@@ -66,12 +66,12 @@ describe('initializeSentryOnEntryClient', () => {
     );
 
     expect(result.generate().code).toMatchInlineSnapshot(`
-      "import * as Sentry from "@sentry/remix";
+      "import {  init, replayIntegration,} from "@sentry/remix";
 
-      Sentry.init({
+      init({
           dsn: "https://sentry.io/123",
 
-          integrations: [Sentry.replayIntegration({
+          integrations: [replayIntegration({
               maskAllText: true,
               blockAllMedia: true
           })],
@@ -106,13 +106,13 @@ describe('initializeSentryOnEntryClient', () => {
         useMatches,
       } from "@remix-run/react";
 
-      import * as Sentry from "@sentry/remix";
+      import {  init, browserTracingIntegration,} from "@sentry/remix";
 
-      Sentry.init({
+      init({
           dsn: "https://sentry.io/123",
           tracesSampleRate: 1,
 
-          integrations: [Sentry.browserTracingIntegration({
+          integrations: [browserTracingIntegration({
             useEffect,
             useLocation,
             useMatches